backend/requirements.txt
Flask
Flask-CORS
requests
python-dotenv

